Fix crash and correctness defects, cut editing and render cost
Full review of src/ across the renderer, text buffers, highlighter,
command layer, memory safety and architecture, with every finding
verified against the code before being acted on.

Bugs

- Fix a use-after-free: `buffer close` destroys the CursorContext that
  Prompt::onKeyDown, ApplicationWindow::mainLoop and ExecCommand keep
  dereferencing afterwards.
- Guard the clipboard UTF-8 conversion in PasteTextCommand. An X11
  Latin-1 selection threw, and main() has no try, so Ctrl+V aborted the
  process and lost every unsaved buffer.
- Disarm a degenerate selection in Cursor::eraseSelection. Shift+Home at
  column 0 left an armed anchor whose stale line index then ran off the
  end of LineBuffer::m_line_data.
- Save atomically via a temporary file and rename, preserving the
  destination's permissions, with an in-place fallback when the
  directory is not writable or rename is unavailable. A failed write no
  longer truncates the previous content.
- Widen AtlasEntry bearings to int16_t and fall back to a blank glyph
  instead of throwing out of render(); font sizes >= 86 aborted with the
  shipped font. Derive the size cap from the face bbox instead of
  hardcoding one font's threshold.
- Throw on the first FT_Request_Size failure. Retaining zero metrics
  divided by zero as soon as a view rendered.
- searchBackward never scanned line 0, so find_prev could not match on
  the first line.
- Snap the cursor off surrogate pairs on vertical moves. Editing at that
  position orphaned a surrogate and made the buffer unsaveable.
- Clamp the View::drawQuad narrowing and the Editor text-area scissor,
  which wrapped or silently disabled clipping at extreme geometry.
- Reject negative goto_line arguments, stop indexing an empty argument
  span in save, compare pending-feedback identity rather than presence,
  and return self-consistent offsets from LineBuffer's empty insert.
- Refresh stick.index on cut and paste. Both omitted it, so the next
  vertical move teleported the cursor to a stale column.

Performance

- Share compiled tree-sitter queries across buffers instead of
  recompiling them per open file: opening 8 C++ buffers 0.88s -> 0.49s.
- Make LineBuffer::insert linear for multi-line input: pasting 10k lines
  into a 40k-line buffer 637ms -> 5.5ms.
- Splice the highlight cache across line-count changes instead of
  dropping a 512-row window, and narrow the undo/redo BufferEdit to the
  differing range (189000 code units -> 1 for a one-character undo).
- Cache search match counts so 98% of find steps skip the full rescan,
  fetch highlight rows once per line rather than once per glyph, resolve
  scrollbar geometry once per frame, and bound undo history by bytes as
  well as by entry count.

Structure

- Extract CursorContext::eraseSelectionIfAny, collapsing 8 duplicated
  call sites whose drift caused the stick.index bug above.
- Extract CommandFeedback::requestArgument/requestPathArgument, folding
  4 sites that carried three divergent path-quoting rules into one.
- Return the batch count from QuadBuffer::endBatch instead of relying on
  staging surviving the call, extract CursorContextManager::makeContext,
  and delete the unused StringBuffer, VectorBuffer and the now-dead
  getHighLightAtPosition.

Verified against the pre-change tree by capturing the rendered window
and pixel-diffing it (0 differing pixels, both static and after a
cut/paste/undo sequence) and by an end-to-end command pass over open,
themes, search, replace, cut, paste, undo, redo, buffer switching, font
size, close and save that produces a byte-identical file. The
stick.index fix is the one intentional behaviour change.

+++ src/core/renderer/AtlasEntry.h
@@ -19,12 +19,18 @@
 #ifndef ATLAS_ENTRY_H
 #define ATLAS_ENTRY_H
 
+#include <cstdint>
+
 
 /**
  * @brief Represents a single entry in the glyph texture atlas.
  *
  * This structure holds texture coordinates, dimensions, and bearing information
  * used for rendering glyphs.
+ *
+ * Coordinates and dimensions are bounded by the atlas texture, which is 255x255x255.
+ * Bearings only offset the quad on screen and never index the texture, so they are
+ * stored wide enough for the values a large font size produces.
  */
 struct AtlasEntry final {
     uint8_t texture_s;    ///< Horizontal starting UV coordinate (S).
@@ -32,8 +38,8 @@ struct AtlasEntry final {
     uint8_t layer;        ///< Layer index within the atlas texture.
     uint8_t width;        ///< Width of the glyph or sprite in pixels.
     uint8_t height;       ///< Height of the glyph or sprite in pixels.
-    int8_t bearing_x;     ///< Horizontal bearing (offset from origin).
-    int8_t bearing_y;     ///< Vertical bearing (offset from baseline).
+    int16_t bearing_x;    ///< Horizontal bearing (offset from origin).
+    int16_t bearing_y;    ///< Vertical bearing (offset from baseline).
 };
